1.过滤规则

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/event/ContactEvtTypeServiceImpl.java
@@ -1,13 +1,16 @@
 package com.zjtelcom.cpct.service.impl.event;
 
 import com.zjtelcom.cpct.constants.CommonConstant;
+import com.zjtelcom.cpct.dao.event.ContactEvtMapper;
 import com.zjtelcom.cpct.dao.event.ContactEvtTypeMapper;
-import com.zjtelcom.cpct.domain.event.EventTypeDO;
+import com.zjtelcom.cpct.dto.event.ContactEvt;
 import com.zjtelcom.cpct.dto.event.ContactEvtType;
 import com.zjtelcom.cpct.dto.event.EventTypeDTO;
 import com.zjtelcom.cpct.request.event.QryContactEvtTypeReq;
 import com.zjtelcom.cpct.service.BaseService;
 import com.zjtelcom.cpct.service.event.ContactEvtTypeService;
+import com.zjtelcom.cpct.util.DateUtil;
+import com.zjtelcom.cpct.util.UserUtil;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -34,13 +37,15 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
 
     @Autowired
     private ContactEvtTypeMapper contactEvtTypeMapper;
+    @Autowired
+    private ContactEvtMapper contactEvtMapper;
 
     /**
      * 查询事件目录
      */
     @Override
-    public  Map<String,Object> qryContactEvtTypeList(QryContactEvtTypeReq qryContactEvtTypeReq) {
-        Map<String,Object> maps = new HashMap<>();
+    public Map<String, Object> qryContactEvtTypeList(QryContactEvtTypeReq qryContactEvtTypeReq) {
+        Map<String, Object> maps = new HashMap<>();
         List<ContactEvtType> contactEvtTypes = new ArrayList<>();
         List<EventTypeDTO> contactEvtTypeList = new ArrayList<>();
         try {
@@ -55,7 +60,7 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
         }
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
-        maps.put("contactEvtTypes",contactEvtTypeList);
+        maps.put("contactEvtTypeList", contactEvtTypeList);
         return maps;
     }
 
@@ -63,18 +68,13 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      * 查询事件目录
      */
     @Override
-    public  Map<String,Object> qryContactEvtTypeLists(QryContactEvtTypeReq qryContactEvtTypeReq) {
-        Map<String,Object> maps = new HashMap<>();
+    public Map<String, Object> qryContactEvtTypeLists(QryContactEvtTypeReq qryContactEvtTypeReq) {
+        Map<String, Object> maps = new HashMap<>();
         List<ContactEvtType> contactEvtTypes = new ArrayList<>();
-        try {
-            contactEvtTypes = contactEvtTypeMapper.qryContactEvtTypeList(qryContactEvtTypeReq);
-        } catch (Exception e) {
-            e.printStackTrace();
-            logger.error("[op:ContactEvtTypeServiceImpl] fail to listEventTypes ", e);
-        }
+        contactEvtTypes = contactEvtTypeMapper.qryContactEvtTypeList(qryContactEvtTypeReq);
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
-        maps.put("contactEvtTypes",contactEvtTypes);
+        maps.put("contactEvtTypes", contactEvtTypes);
         return maps;
     }
 
@@ -82,16 +82,21 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      * 新增事件目录保存
      */
     @Override
-    public Map<String,Object> createContactEvtType(ContactEvtType contactEvtType) {
-        Map<String,Object> maps = new HashMap<>();
-        try {
-            contactEvtTypeMapper.createContactEvtType(contactEvtType);
-        } catch (Exception e) {
-            e.printStackTrace();
-            logger.error("[op:ContactEvtTypeServiceImpl] fail to createContactEvtType ", e);
-        }
+    public Map<String, Object> createContactEvtType(ContactEvtType contactEvtType) {
+        Map<String, Object> maps = new HashMap<>();
+        contactEvtType.setCreateDate(DateUtil.getCurrentTime());
+        contactEvtType.setUpdateDate(DateUtil.getCurrentTime());
+        contactEvtType.setStatusDate(DateUtil.getCurrentTime());
+        contactEvtType.setUpdateStaff(UserUtil.loginId());
+        contactEvtType.setCreateStaff(UserUtil.loginId());
+        contactEvtType.setStatusCd(CommonConstant.STATUSCD_EFFECTIVE);
+        contactEvtType.setParEvtTypeId(contactEvtType.getEvtTypeId());
+        contactEvtType.setEvtTypeId(EVT_TYPE_ID_NULL);
+        contactEvtTypeMapper.createContactEvtType(contactEvtType);
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
+        maps.put("contactEvtType", contactEvtType);
+        maps.put("ruleEvents", null);
         return maps;
     }
 
@@ -99,8 +104,8 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      * 查看事件目录
      */
     @Override
-    public Map<String,Object> getEventTypeDTOById(Long evtTypeId) {
-        Map<String,Object> maps = new HashMap<>();
+    public Map<String, Object> getEventTypeDTOById(Long evtTypeId) {
+        Map<String, Object> maps = new HashMap<>();
         ContactEvtType contactEvtType = new ContactEvtType();
         try {
             contactEvtType = contactEvtTypeMapper.selectByPrimaryKey(evtTypeId);
@@ -110,7 +115,7 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
         }
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
-        maps.put("contactEvtType",contactEvtType);
+        maps.put("contactEvtType", contactEvtType);
         return maps;
     }
 
@@ -118,16 +123,15 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      * 修改事件目录
      */
     @Override
-    public Map<String,Object> modContactEvtType(ContactEvtType contactEvtType) {
-        Map<String,Object> maps = new HashMap<>();
-        try {
-            contactEvtTypeMapper.modContactEvtType(contactEvtType);
-        } catch (Exception e) {
-            e.printStackTrace();
-            logger.error("[op:ContactEvtTypeServiceImpl] fail to modContactEvtType ", e);
-        }
+    public Map<String, Object> modContactEvtType(ContactEvtType contactEvtType) {
+        Map<String, Object> maps = new HashMap<>();
+        contactEvtType.setUpdateDate(DateUtil.getCurrentTime());
+        contactEvtType.setUpdateStaff(UserUtil.loginId());
+        contactEvtTypeMapper.modContactEvtType(contactEvtType);
         maps.put("resultCode", CommonConstant.CODE_SUCCESS);
         maps.put("resultMsg", StringUtils.EMPTY);
+        maps.put("contactEvtType", contactEvtType);
+        maps.put("ruleEvents", null);
         return maps;
     }
 
@@ -135,16 +139,28 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      * 删除事件目录
      */
     @Override
-    public Map<String,Object> delContactEvtType(ContactEvtType contactEvtType) {
-        Map<String,Object> maps = new HashMap<>();
-        try {
+    public Map<String, Object> delContactEvtType(ContactEvtType contactEvtType) {
+        Map<String, Object> maps = new HashMap<>();
+        ContactEvt contactEvt = new ContactEvt();
+        QryContactEvtTypeReq qryContactEvtTypeReq = new QryContactEvtTypeReq();
+        qryContactEvtTypeReq.setParEvtTypeId(contactEvtType.getEvtTypeId());
+        contactEvt.setContactEvtTypeId(contactEvtType.getEvtTypeId());
+        List<ContactEvt> contactEvtList = contactEvtMapper.listEvents(contactEvt);
+        List<ContactEvtType> contactEvtTypeList = contactEvtTypeMapper.qryContactEvtTypeList(qryContactEvtTypeReq);
+        if (contactEvtList.size() > 0) {
+            maps.put("resultCode", CommonConstant.CODE_FAIL);
+            maps.put("resultMsg", "事件类型已关联事件，不可删除！");
+            return maps;
+        } else if (contactEvtTypeList.size() > 0) {
+            maps.put("resultCode", CommonConstant.CODE_FAIL);
+            maps.put("resultMsg", "事件类型有子类，不可删除！");
+            return maps;
+        } else {
             contactEvtTypeMapper.deleteByPrimaryKey(contactEvtType.getEvtTypeId());
-        } catch (Exception e) {
-            e.printStackTrace();
-            logger.error("[op:ContactEvtTypeServiceImpl] fail to delContactEvtType ", e);
+            maps.put("resultCode", CommonConstant.CODE_SUCCESS);
+            maps.put("resultMsg", StringUtils.EMPTY);
+            maps.put("ruleEvents", null);
         }
-        maps.put("resultCode", CommonConstant.CODE_SUCCESS);
-        maps.put("resultMsg", StringUtils.EMPTY);
         return maps;
     }
 
@@ -153,37 +169,36 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
      */
     private List<EventTypeDTO> generateTree(List<ContactEvtType> contactEvtTypes) {
         List<EventTypeDTO> dtoList = new ArrayList<>();
-        List<EventTypeDTO> eventTypeDTOS = new ArrayList<>();
         for (ContactEvtType contactEvtType : contactEvtTypes) {
             if (contactEvtType.getParEvtTypeId() == PAR_EVT_TYPE_ID_ZERO) {
                 EventTypeDTO eventTypeDTO = new EventTypeDTO();
-                eventTypeDTO.setEvtTypeId(contactEvtType.getEvtTypeId());
-                eventTypeDTO.setParEvtTypeId(contactEvtType.getParEvtTypeId());
                 eventTypeDTO.setContactEvtName(contactEvtType.getContactEvtName());
+                eventTypeDTO.setParEvtTypeId(contactEvtType.getParEvtTypeId());
+                eventTypeDTO.setEvtTypeId(contactEvtType.getEvtTypeId());
                 dtoList.add(eventTypeDTO);
             }
             // 为一级菜单设置子菜单，getChild是递归调用的
             for (EventTypeDTO eventTypeDTO : dtoList) {
-                eventTypeDTO.setEventTypeDTOList(getChild(eventTypeDTO.getEvtTypeId(), eventTypeDTOS));
+                eventTypeDTO.setChildren(getChild(eventTypeDTO.getEvtTypeId(), contactEvtTypes));
             }
         }
-        return eventTypeDTOS;
+        return dtoList;
     }
 
     /**
      * 递归查找子菜单
      */
-    private List<EventTypeDTO> getChild(Long id, List<EventTypeDTO> rootMenu) {
+    private List<EventTypeDTO> getChild(Long id, List<ContactEvtType> rootMenu) {
         // 子菜单
         List<EventTypeDTO> childList = new ArrayList<>();
-        for (EventTypeDTO eventTypeDTO : rootMenu) {
+        for (ContactEvtType eventTypeDTO : rootMenu) {
             // 遍历所有节点，将父菜单id与传过来的id比较
             if (eventTypeDTO.getParEvtTypeId() != PAR_EVT_TYPE_ID_ZERO) {
                 if (eventTypeDTO.getParEvtTypeId() == id) {
                     EventTypeDTO eventTypeDTOS = new EventTypeDTO();
-                    eventTypeDTO.setContactEvtName(eventTypeDTO.getContactEvtName());
-                    eventTypeDTO.setEvtTypeId(eventTypeDTO.getEvtTypeId());
-                    eventTypeDTO.setParEvtTypeId(eventTypeDTO.getParEvtTypeId());
+                    eventTypeDTOS.setContactEvtName(eventTypeDTO.getContactEvtName());
+                    eventTypeDTOS.setEvtTypeId(eventTypeDTO.getEvtTypeId());
+                    eventTypeDTOS.setParEvtTypeId(eventTypeDTO.getParEvtTypeId());
                     childList.add(eventTypeDTOS);
                 }
             }
@@ -196,7 +211,7 @@ public class ContactEvtTypeServiceImpl extends BaseService implements ContactEvt
             List<ContactEvtType> list = contactEvtTypeMapper.qryContactEvtTypeList(qryContactEvtTypeReq);
             if (list.size() != LIST_SIZE_ZERO) {
                 // 递归
-                eventTypeDTO.setEventTypeDTOList(getChild(eventTypeDTO.getEvtTypeId(), rootMenu));
+                eventTypeDTO.setChildren(getChild(eventTypeDTO.getEvtTypeId(), rootMenu));
             }
         } // 递归退出条件
         if (childList.size() == LIST_SIZE_ZERO) {
